few bugfixes. print_grammar (--print-grammar, -pg) now recognizes all node types.
git-svn-id: https://tinyap.googlecode.com/svn/trunk@40 26adf5cf-fd31-0410-b5e7-cb6bf36da140

+++ src/tokenizer.c
@@ -920,7 +920,7 @@ const char* parse_error(token_context_t*t) {
 
 	sep = " In context ";
 
-	for(i=0;i<=t->farthest_stack->sp;i+=1) {
+	for(i=0;i<=(long)t->farthest_stack->sp;i+=1) {
 		k=(char*)t->farthest_stack->stack[i];
 		if(k) {
 			strcat(err_buf,sep);
